requirements.txt
newsapi-python
pymongo
pyspark
kafka-python